requirements.txt
requests
groq
python-dotenv
sentence_transformers
faiss-cpu
flask-cors
flask